refactor: change orderTable classes structure by BOUNDED CONTEXT
